# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "FrontendCreatePatient": {
        "properties": {
          "bed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bed"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "repositioningInterval": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repositioninginterval"
          },
          "riskLevel": {
            "title": "Risklevel",
            "type": "string"
          },
          "room": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Room"
          }
        },
        "required": [
          "name",
          "riskLevel"
        ],
        "title": "FrontendCreatePatient",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "PacienteConfigResponse": {
        "properties": {
          "cama_id": {
            "title": "Cama Id",
            "type": "string"
          },
          "nome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nome"
          },
          "observacoes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observacoes"
          },
          "paciente_id": {
            "title": "Paciente Id",
            "type": "string"
          },
          "perfil": {
            "title": "Perfil",
            "type": "string"
          },
          "rotinas": {
            "items": {
              "$ref": "#/components/schemas/RotinaConfig"
            },
            "title": "Rotinas",
            "type": "array"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "paciente_id",
          "cama_id",
          "perfil",
          "rotinas"
        ],
        "title": "PacienteConfigResponse",
        "type": "object"
      },
...
      "TrocaDeLeitosRequest": {
        "properties": {
          "pacienteA": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Pacientea",
            "type": "string"
          },
          "pacienteB": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Pacienteb",
            "type": "string"
          }
        },
        "required": [
          "pacienteA",
          "pacienteB"
        ],
        "title": "TrocaDeLeitosRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Monitor de Alertas UPP",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/pacientes": {
      "get": {
        "description": "Pacientes internados nas unidades que o chamador enxerga.\n\n`incluir_alta` traz tambem os que ja sairam. Fora dele a lista e a ala de\nagora: depois que alta virou estado (migrations/0009), incluir quem foi\nembora encheria a tela de gente que nao esta mais no predio.",
        "operationId": "listar_pacientes_endpoint_api_pacientes_get",
        "parameters": [
          {
            "in": "query",
            "name": "incluir_alta",
            "required": false,
            "schema": {
              "default": false,
              "title": "Incluir Alta",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Listar Pacientes Endpoint Api Pacientes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Listar Pacientes Endpoint",
        "tags": [
          "pacientes"
        ]
      },
      "post": {
        "description": "Criar um novo paciente.",
        "operationId": "criar_paciente_endpoint_api_pacientes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FrontendCreatePatient"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Criar Paciente Endpoint Api Pacientes Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Criar Paciente Endpoint",
        "tags": [
          "pacientes"
        ]
      }
    },
    "/api/pacientes/cama/{cama_id}": {
      "get": {
        "operationId": "obter_paciente_por_cama_endpoint_api_pacientes_cama__cama_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "cama_id",
            "required": true,
            "schema": {
              "title": "Cama Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PacienteConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Obter Paciente Por Cama Endpoint",
        "tags": [
          "pacientes"
        ]
      }
    },
    "/api/pacientes/troca-de-leitos": {
      "post": {
        "description": "Dois pacientes trocam de leito.\n\nEra impossivel pela tela: qualquer sequencia de duas edicoes passa por um\nestado em que um leito tem dois ocupantes, e o indice unico parcial de\n`cama_id` recusa. Trocar dois pacientes de lugar e rotina numa ala.",
        "operationId": "trocar_leitos_endpoint_api_pacientes_troca_de_leitos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrocaDeLeitosRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Trocar Leitos Endpoint Api Pacientes Troca De Leitos Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trocar Leitos Endpoint",
        "tags": [
          "pacientes"
        ]
      }
    },
    "/api/pacientes/{paciente_id}": {
      "delete": {
        "description": "Remove um paciente e todo o rastro clinico dele.\n\nNAO e o caminho para dar alta — para isso existe\n`POST /api/pacientes/{id}/alta`, que preserva o historico. Este endpoint\ne para ERRO DE CADASTRO (paciente duplicado, criado no lugar errado), que e\ncoisa diferente, rara, e por isso segue restrito a admin.\n\nA rota nao existia. O `PatientRepository.delete` e o `dao.remover_paciente`\nja estavam escritos e sem nenhum chamador, e a tela de Pacientes tinha o\nbotao \"Excluir\" com dialogo de confirmacao chamando `DELETE /api/pacientes/\n{id}` — que respondia 405. O usuario confirmava a exclusao, via \"Erro ao\nremover paciente\" e o paciente continuava na lista.\n\nRestrito a admin, pelo mesmo motivo de `/simular`: e uma acao irreversivel\nsobre o historico clinico de uma pessoa. O acesso ja entra na trilha de\nauditoria pelo middleware (a rota casa com `/api/pacientes`), entao fica\nregistrado QUEM removeu e QUANDO.",
        "operationId": "remover_paciente_endpoint_api_pacientes__paciente_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "paciente_id",
            "required": true,
            "schema": {
              "title": "Paciente Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Remover Paciente Endpoint Api Pacientes  Paciente Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remover Paciente Endpoint",
        "tags": [
          "pacientes"
        ]
      },
      "get": {
        "description": "Obter um paciente pelo ID.",
        "operationId": "obter_paciente_endpoint_api_pacientes__paciente_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "paciente_id",
            "required": true,
            "schema": {
              "title": "Paciente Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Obter Paciente Endpoint Api Pacientes  Paciente Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Obter Paciente Endpoint",
        "tags": [
          "pacientes"
        ]
      },
      "patch": {
        "description": "Atualizar um paciente existente.",
        "operationId": "atualizar_paciente_endpoint_api_pacientes__paciente_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "paciente_id",
            "required": true,
            "schema": {
              "title": "Paciente Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FrontendCreatePatient"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Atualizar Paciente Endpoint Api Pacientes  Paciente Id  Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Atualizar Paciente Endpoint",
        "tags": [
          "pacientes"
        ]
      }
    },
...
    "/api/pacientes/{paciente_id}/alta": {
      "post": {
        "description": "Encerra a internacao preservando o historico clinico.\n\nExiste porque a unica forma de tirar um paciente da tela era `DELETE`, que\napaga alertas, grade, eventos e timeline. A operacao mais rotineira de uma\nala destruia a evidencia que acreditacao e LGPD exigem guardar.\n\nNao e restrito a admin, diferente do `DELETE`: dar alta e ato clinico de\nrotina da equipe, e exigir admin para isso empurraria a enfermagem de volta\npara o botao de excluir — que e destrutivo.",
        "operationId": "dar_alta_endpoint_api_pacientes__paciente_id__alta_post",
        "parameters": [
          {
            "in": "path",
            "name": "paciente_id",
            "required": true,
            "schema": {
              "title": "Paciente Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AltaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Dar Alta Endpoint Api Pacientes  Paciente Id  Alta Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dar Alta Endpoint",
        "tags": [
          "pacientes"
        ]
      }
    },
    "/api/pacientes/{paciente_id}/simular": {
      "post": {
        "description": "Simular dados históricos para um paciente.\n\nRestrito a admin: grava eventos, grade e ALERTAS sintéticos no mesmo banco\ndos dados reais. Num ambiente de produção, dado simulado misturado ao\nhistórico clínico do paciente é indistinguível do que veio do sensor.",
        "operationId": "simular_paciente_endpoint_api_pacientes__paciente_id__simular_post",
        "parameters": [
          {
            "in": "path",
            "name": "paciente_id",
            "required": true,
            "schema": {
              "title": "Paciente Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Simular Paciente Endpoint",
        "tags": [
          "pacientes"
        ]
      }
    },
    "/api/pacientes/{paciente_id}/transferencia": {
      "post": {
        "description": "Move o paciente de leito como operacao propria.\n\nAntes era efeito colateral de editar um campo do formulario, o que perdia\nduas coisas: a transferencia E um reposicionamento (ser erguido para a maca\ne alivio de pressao real), e o estado do motor seguia o paciente com o\nrelogio e a corrida do leito anterior.",
        "operationId": "transferir_endpoint_api_pacientes__paciente_id__transferencia_post",
        "parameters": [
          {
            "in": "path",
            "name": "paciente_id",
            "required": true,
            "schema": {
              "title": "Paciente Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferenciaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Transferir Endpoint Api Pacientes  Paciente Id  Transferencia Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Transferir Endpoint",
        "tags": [
          "pacientes"
        ]
      }
    },
...
  }
}